Wshadow

+++ src/lgfx/v1/panel/Panel_M5HDMI.cpp
@@ -44,7 +44,7 @@ namespace lgfx
 //----------------------------------------------------------------------------
 
   enum GWFPGA_Inst_Def
-  { 
+  {
     ISC_NOOP          = 0x02,
     ISC_ERASE         = 0x05,
     ERASE_DONE        = 0x09,
@@ -124,7 +124,7 @@ namespace lgfx
     ESP_LOGI(TAG, "Starting Writing to SRAM...");
     JTAG_WriteInst(ISC_ENABLE);
     JTAG_WriteInst(FAST_PROGRAM);
-    
+
     JTAG_MoveTap(TAP_IDLE, TAP_DRSHIFT);
 
     int32_t rle_len = -1;
@@ -286,7 +286,7 @@ namespace lgfx
       if (lgfx::gpio_in(TDO_PIN)) { out += 1 << i; }
     };
     JTAG_MoveTap(TAP_DREXIT1,  TAP_IDLE);
-    return out;   
+    return out;
   }
 
 //----------------------------------------------------------------------------
@@ -451,7 +451,7 @@ namespace lgfx
     }
 
     if (!Panel_Device::init(false)) { return false; }
-  
+
     // Initialize and read ID
     ESP_LOGI(TAG, "Waiting the FPGA gets idle...");
     startWrite();
@@ -470,7 +470,7 @@ namespace lgfx
     cs_control(true);
     _bus->endRead();
     cs_control(false);
-  
+
     bool res = _init_resolution();
 
     ESP_LOGI(TAG, "Initialize HDMI transmitter...");
@@ -517,9 +517,9 @@ namespace lgfx
 
     bool res = (hori_total > hori_min);
     if (!res)
-    { // If the blanking period is too small, it will not work properly. 
+    { // If the blanking period is too small, it will not work properly.
       hori_total = hori_min;
-      ESP_LOGE(TAG, "resolution error. out of range  %dx%d %.2f Hz", mem_width, mem_height, _refresh_rate);
+      ESP_LOGE(TAG, "resolution error. out of range  %dx%d %.2f Hz", mem_width, mem_height, (double)_refresh_rate);
     }
 
     video_timing_t vt;
